refactor(config): move configuration logic to util/settings.py for better modularity refactor(main.py): update imports and usage to reflect new module structure feat(util): add replicate_api.py to encapsulate API interactions refactor(util): create settings.py to handle configuration management
